fix: satisfy structured finality protocol checks
