+++ apps/shared/OpenClawKit/Sources/OpenClawProtocol/GatewayModels.swift
@@ -13972,6 +13972,7 @@ public struct CronRunLogEntry: Codable, Sendable {
     public let error: String?
     public let errorreason: AnyCodable?
     public let summary: String?
+    public let assistantcompletion: [String: AnyCodable]?
     public let diagnostics: [String: AnyCodable]?
     public let delivered: Bool?
     public let deliverystatus: AnyCodable?
@@ -13997,6 +13998,7 @@ public struct CronRunLogEntry: Codable, Sendable {
         error: String? = nil,
         errorreason: AnyCodable? = nil,
         summary: String? = nil,
+        assistantcompletion: [String: AnyCodable]? = nil,
         diagnostics: [String: AnyCodable]? = nil,
         delivered: Bool? = nil,
         deliverystatus: AnyCodable? = nil,
@@ -14021,6 +14023,7 @@ public struct CronRunLogEntry: Codable, Sendable {
         self.error = error
         self.errorreason = errorreason
         self.summary = summary
+        self.assistantcompletion = assistantcompletion
         self.diagnostics = diagnostics
         self.delivered = delivered
         self.deliverystatus = deliverystatus
@@ -14047,6 +14050,7 @@ public struct CronRunLogEntry: Codable, Sendable {
         case error
         case errorreason = "errorReason"
         case summary
+        case assistantcompletion = "assistantCompletion"
         case diagnostics
         case delivered
         case deliverystatus = "deliveryStatus"
